Removes variable length arrays and uses fixed length of N
This is the only way to work with Visual Studio.
VS has an issue due to being run on the .net framework
where variable length arrays cannnot be alocated, due to
.net allocating them directly on the heap. The traditional
method to get around this is to use the C new operator.
As I have made a C++11 program, this is innapropriate
and I will instead sacrifice memory to gain cross compatibility
with windows 7. This is no longer an issue in Windows 10 as
in 10 you have access to the bash terminal.

Also, begins writing more of the homework.